# docs/apispecs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kasir App",
        "description": "Kasir App API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/v1"
        }
    ],
    "paths": {
...
        "/suppliers/{supplier_id}": {
            "put": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Update Supplier",
                "summary": "Update Supplier",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "pattern": "^08[0-9]{9,}$"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Supplier Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/suppliers_response"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Suppliers API"
                ],
                "description": "Delete Supplier",
                "summary": "Delete Supplier",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/discounts": {
            "get": {
                "tags": [
                    "Discounts API"
                ],
                "description": "Get Discounts",
                "summary": "Get Discounts",
                "operationId": "getDiscounts",
                "responses": {
                    "200": {
                        "description": "Discounts Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/discounts_response"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Discounts API"
                ],
                "description": "Create Discount",
                "summary": "Create Discount",
                "operationId": "createDiscount",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Discount Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/discounts_response"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/discounts/{discount_id}": {
            "get": {
                "tags": [
                    "Discounts API"
                ],
                "description": "Get Discount",
                "summary": "Get Discount",
                "operationId": "getDiscount",
                "parameters": [
                    {
                        "name": "discount_id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/discounts_response"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Discounts API"
                ],
                "description": "Delete Discount",
                "summary": "Delete Discount",
                "operationId": "deleteDiscount",
                "parameters": [
                    {
                        "name": "discount_id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Discounts API"
                ],
                "description": "Update Discount",
                "summary": "Update Discount",
                "operationId": "updateDiscount",
                "parameters": [
                    {
                        "name": "discount_id",
                        "in": "path",
                        "description": "Discount ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "percentage": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discount Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/discounts_response"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
            "schemas": {
...
                "suppliers_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "full_name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "phone": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
                "discounts_response": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number"
                        },
                        "name": {
                            "type": "string"
                        },
                        "percentage": {
                            "type": "number"
                        },
                        "created_at": {
                            "type": "string"
                        }
                    }
                },
...
            },
            "securitySchemes": {
                    "BearerAuth": {
                        "type": "http",
                        "scheme": "bearer",
                        "bearerFormat": "JWT"
                    }
                }
        }
}